Reshuffle multiprocess module
