+++ src/sentry/sentry_metrics/consumers/last_seen_updater.py
@@ -14,9 +14,9 @@ from arroyo.processing.strategies.streaming.factory import StreamMessageFilter
 from django.utils import timezone
 
 from sentry import options
+from sentry.sentry_metrics.consumers.indexer.multiprocess import get_config, logger
 from sentry.sentry_metrics.indexer.base import FetchType
 from sentry.sentry_metrics.indexer.models import StringIndexer
-from sentry.sentry_metrics.multiprocess import get_config, logger
 from sentry.utils import json
 
 MAPPING_META = "mapping_meta"
